Update 5h.html
In line No. 12 it should be pre increment instead of post increment or it is not showing the correct result when we are clicking the button by using post increment operator.
In line No. 22 ,I have used the addition assignment operator to make the code small and identical to other addition lines.